OT/PLC/requirements.txt
python-snap7
paho-mqtt
influxdb-client
